+++ CHANGES.rst
@@ -31,6 +31,8 @@ Improvements
   delete Indico users from the user management UI (:pr:`6652`, thanks :user:`SegiNyn`)
 - Support ``==text==`` to highlight text in markdown (:issue:`6731`, :pr:`6732`)
 - Replace the "Disallow custom persons" event setting with "Enforce user search" (:pr:`6689`)
+- Add an event setting to allow enforcing search before entering a person manually to
+  a persons list in abstracts and contributions (:pr:`6689`)
 
 Bugfixes
 ^^^^^^^^